Add compat modules for `tomllib` and `importlib.metadata`
