+++ piptools/_compat/tomllib.py
@@ -1,0 +1,11 @@
+from __future__ import annotations
+
+import sys
+
+if sys.version_info >= (3, 11):
+    from tomllib import TOMLDecodeError, load, loads
+else:
+    from tomli import TOMLDecodeError, load, loads
+
+
+__all__ = ["loads", "load", "TOMLDecodeError"]
